Get Content body name.

Get naver news body from link.

getContentBody(
  html_obj,
  body_node_info = "div#articleBodyContents",
  body_attr = ""
)

Arguments

html_obj
"xml_document" "xml_node" using read_html function.

body_node_info
Information about node names like tag with class or id. Default is "div.article_info h3" for naver news title.

body_attr
if you want to get attribution text, please write down here.

Value

Get character body content.

Examples

# \donttest{
  print(news_url_ex)

#> [1] "https://news.naver.com/main/read.nhn?mode=LSD&mid=shm&sid1=101&oid=374&aid=0000184049"

  hobj <- xml2::read_html(news_url_ex)
  getContentBody(hobj)

#> [1] "동영상 뉴스 \t\t \t    서울 3곳, 인천 1곳, 광주 1곳, 충남 1곳…출혈·과당경쟁 우려[앵커]정부가 시내면세점 추가 특허 수를 결정했습니다.취재기자 연결합니다.장가희 기자, 정부가 허가를 내주기로 한 신규 시내면세점 얼마나 됩니까?[기자]기획재정부는 오늘(14일) 오후 보세판매장 제도운영위원회를 열고, 신규 시내 면세점을 6곳 추가로 출점할 것을 결정했습니다.대기업 시내면세점은 서울 3곳, 인천 1곳, 광주 1곳이며 중소, 중견기업은 충남 1곳입니다.서울의 경우 올해는 총 특허 수 제한 없이 개별기업 신청에 의해 심사 후 특허를 부여한다는 방침입니다.기재부는 위원회 심의 결과를 관세청에 통보하고, 관세청은 오는 11월 최종 사업자 선정을 할 예정입니다.[앵커]그런데 장기자, 가뜩이나 시내면세점 부진을 겪고 있는데, 또 면세점이 들어서면, 출혈경쟁이 더 심해지는 게 아닌가요?[기자]네, 국내 면세시장은 매월 매출 신기록을 내고 있지만, 중국인 보따리상 등을 유치하기 위해 과도한 수수료 경쟁으로 면세점에도 부익부 빈익빈이 심해진 상태입니다.여기에 최근 한화가 면세사업을 포기하며 그동안 적자에도 불구하고 면세사업을 강행하던 다른 면세점에도 파장이 예상됩니다.여기에 정부가 입국장 면세점까지 신규 사업 허가를 내주면서 이미 포화상태인 면세시장이 침체될 것이란 우려도 나옵니다.지금까지 SBSCNBC 장가희입니다.장가희 기자(gani@sbs.co.kr)☞ 의견 있는 경제채널 SBSCNBC [홈페이지]☞ 경제를 실험한다~ #머니랩 [네이버 TV] 저작권자 SBS플러스 & SBS I&M 무단전재-재배포 금지"

  # }