Get Content datetime

Get naver news published datetime from link.

getContentDatetime(
  html_obj,
  datetime_node_info = "span.t11",
  datetime_attr = "",
  getEdittime = TRUE
)

Arguments

html_obj
"xml_document" "xml_node" using read_html function.

datetime_node_info
Information about node names like tag with class or id. Default is "div.article_info h3" for naver news title.

datetime_attr
if you want to get attribution text, please write down here.

getEdittime
if TRUE, can get POSIXlt type datetime length 2 means published time and final edited time. if FALSE, get Date length 1.

Value

Get POSIXlt type datetime.

Examples

# \donttest{
  print(news_url_ex)
#> [1] "https://news.naver.com/main/read.nhn?mode=LSD&mid=shm&sid1=101&oid=374&aid=0000184049"
  hobj <- xml2::read_html(news_url_ex)
  getContentDatetime(hobj)
#> [1] "2019-05-14 21:04:00 KST" "2019-05-14 21:52:00 KST"
  # }