Get All Comment

Get all comments from the provided news article url on naver

getAllComment(turl = url, ...)

Arguments

turl
character. News article on 'Naver' such as <http://news.naver.com/main/read.nhn?mode=LSD&mid=shm&sid1=100&oid=056&aid=0010335895>. News articl url that is not on Naver.com domain will generate an error.

...
parameter in getComment function.

Value

a [tibble][tibble::tibble-package]

Details

Works just like getComment, but this function executed in a fashion where it finds and extracts all comments from the given url.

Examples

  print(news_url_ex)

#> [1] "https://news.naver.com/main/read.nhn?mode=LSD&mid=shm&sid1=101&oid=374&aid=0000184049"

  getAllComment(news_url_ex)

#> # A tibble: 160 x 76
#>    ticket objectId categoryId templateId commentNo parentCommentNo replyLevel
#>    <chr>  <chr>    <chr>      <chr>          <int>           <int>      <int>
#>  1 news   news374… *          default_e…    1.71e9      1707777412          1
#>  2 news   news374… *          view_econ…    1.71e9      1707772282          1
#>  3 news   news374… *          view_econ…    1.71e9      1707752242          1
#>  4 news   news374… *          default_e…    1.71e9      1707723522          1
#>  5 news   news374… *          view_econ…    1.71e9      1707710742          1
#>  6 news   news374… *          view_econ…    1.71e9      1707691422          1
#>  7 news   news374… *          view_econ…    1.71e9      1707682112          1
#>  8 news   news374… *          view_econ…    1.71e9      1707645262          1
#>  9 news   news374… *          view_econ…    1.71e9      1707644392          1
#> 10 news   news374… *          default_e…    1.71e9      1707638522          1
#> # … with 150 more rows, and 69 more variables: replyCount <int>,
#> #   replyAllCount <int>, replyPreviewNo <lgl>, replyList <lgl>,
#> #   imageCount <int>, imageList <lgl>, imagePathList <lgl>,
#> #   imageWidthList <lgl>, imageHeightList <lgl>, commentType <chr>,
#> #   stickerId <lgl>, sticker <lgl>, sortValue <dbl>, contents <chr>,
#> #   userIdNo <chr>, exposedUserIp <lgl>, lang <chr>, country <chr>,
#> #   idType <chr>, idProvider <chr>, userName <chr>, userProfileImage <chr>,
#> #   profileType <chr>, modTime <chr>, modTimeGmt <chr>, regTime <chr>,
#> #   regTimeGmt <chr>, sympathyCount <int>, antipathyCount <int>,
#> #   userBlind <lgl>, hideReplyButton <lgl>, status <int>, mine <lgl>,
#> #   best <lgl>, mentions <lgl>, toUser <lgl>, userStatus <int>,
#> #   categoryImage <lgl>, open <lgl>, levelCode <lgl>, grades <lgl>,
#> #   sympathy <lgl>, antipathy <lgl>, metaInfo <lgl>, extension <lgl>,
#> #   audioInfoList <lgl>, translation <lgl>, report <lgl>,
#> #   middleBlindReport <lgl>, spamInfo <lgl>, userHomepageUrl <lgl>,
#> #   defamation <lgl>, hiddenByCleanbot <lgl>, visible <lgl>, serviceId <lgl>,
#> #   deleted <lgl>, secret <lgl>, blind <lgl>, manager <lgl>, anonymous <lgl>,
#> #   blindReport <lgl>, expose <lgl>, profileUserId <lgl>,
#> #   exposeByCountry <lgl>, virtual <lgl>, containText <lgl>,
#> #   maskedUserId <chr>, maskedUserName <chr>, validateBanWords <lgl>